+++ src/constants/alCommands.ts
@@ -1,3 +1,4 @@
 export const debugPublish = "dw-vs.al_debugPublish";
 export const createPermissionSet = "dw-vs.al_createPermissionSet";
 export const insertAffix = "dw-vs.al_insertAffix";
+export const renameFiles = "dw-vs.al_renameFiles";