+++ src/constants/alDisplayResources.ts
@@ -4,3 +4,4 @@ export const schemaUpdateMethodQst =
   "One or more of your launch configurations is set to 'Recreate' and data could be lost on publishing if you proceed!";
 export const continueAction = "Continue";
 export const configurationsAction = "View Configurations";
+export const filesRenamed = "files have been renamed.";