+++ src/constants/alPermissionResources.ts
@@ -3,6 +3,8 @@ export const permissionSetFileEnd = "permissionSet.al";
 export const defaultPermissionSetName = "Permission";
 export const permissionTable = "rimd";
 export const permissionExecute = "x";
+export const permissionFileCreated =
+  "Permission set file has been created successfully.";
 
 export const permissionObjectTypeRegEx =
   /\b(CODEUNIT|PAGE|TABLE|REPORT|XMLPORT)\b/i;