+++ src/helpers/textHelper.ts
@@ -5,6 +5,10 @@ export class TextHelper {
     return value.replace(/(\s)/, "");
   }
 
+  stripInvalidChars(value: string): string {
+    return value.replace(/([^\s\w])/g, "");
+  }
+
   toCamelCase(value: string): string {
     return value.replace(/(?:^\w|[A-Z]|\b\w|\s+)/g, function (match, index) {
       if (+match === 0) {